+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Apply `compatibleProvider`,`namespace` metadata validation based on the relevant `AppCatalogEntry` CR.
 
+### Fixed
+
+- Don't default `.spec.config` if app is a Management Cluster app.
+
 ## [0.5.1] - 2021-02-17
 
 ### Fixed